+++ commands/owner/evaluate.js
@@ -5,12 +5,13 @@ executed by the bot owner defined in the config/bot/settings.json file.
 
 const { ownerID } = require("../../config/bot/settings.json");
 const { inspect } = require("util");
+const fs = require("fs");
 
 module.exports = {
     config: {
         name: "evaluate",
         description: "Evaluates JavaScript and Discord.js code. Restricted to the bot owner.",
-        usage: "<input>",
+        usage: ["<input>"],
         category: "owner",
         aliases: ["eval", "js"]
     },
@@ -22,18 +23,24 @@ module.exports = {
           return message.channel.send(`**${message.author.username}**, please enter some code to evaluate!`);
         }
 
-        let toEval = args.join(" ");
+        let toEval = args.join(" ").toLowerCase();
         try {
         	let evaluated = inspect(eval(toEval, { depth: 0 }));
 
           let hrStart = process.hrtime();
           let hrDiff = process.hrtime(hrStart);
-          if (evaluated.length < 2000) {
-            return message.channel.send(`\`\`\`javascript\n${evaluated}\n\`\`\`*(Executed in ${hrDiff[0] > 0 ? `${hrDiff[0]}s ` : ''}${(hrDiff[1] / 1000000).toFixed(3)}ms.)*`, { maxLength: 2000 });
+          if (evaluated.length < 1950) {
+            return message.channel.send(`\`\`\`javascript\n${evaluated}\n\`\`\`*(Executed in ${hrDiff[0] > 0 ? `${hrDiff[0]}s ` : ''}${(hrDiff[1] / 1000000).toFixed(3)}ms.)*`);
           }
           else {
-            console.log(evaluated);
-            return message.channel.send("The evaluated code is longer than 2000 characters!\n(I can't show it here, but the result is in the logs.)");
+            fs.writeFile("./config/bot/eval.txt", evaluated, 'utf8', (err) => {
+              if (err) {
+                console.log(err);
+                return message.channel.send(`Sorry **${message.author.username}**, I couldn't write the evaluated code to file!`);
+              }
+              console.log(evaluated);
+              return message.channel.send("The evaluated code is longer than Discord's message limit, but the result is in the console's logs. In case you can't access that now, I made a text file containing the same data.", { files: ["./config/bot/eval.txt"] } );
+            });
           }
         }
         catch (e) {